- Huge memory leak in Label removed - Texture name (GL name that is) init to 0 - Closed Cardial Cubic splines - Modulo function that works

+++ Furiosity/Collisions/CollisionManager.cpp
@@ -282,9 +282,8 @@ void CollisionManager::AccumulateContacts(const std::list<Entity2D*>& entities)
         }
     }
     
-    next = contacts.size();
-        
-    
+    next =  static_cast<int>(contacts.size());
+            
     
     /*
     // Go over all entities
